+++ src/backend/InvenTree/part/serializers.py
@@ -599,6 +599,7 @@ class PartSerializer(
             'link',
             'locked',
             'minimum_stock',
+            'maximum_stock',
             'name',
             'notes',
             'parameters',
@@ -898,6 +899,10 @@ class PartSerializer(
 
     minimum_stock = serializers.FloatField(
         required=False, label=_('Minimum Stock'), default=0
+    )
+
+    maximum_stock = serializers.FloatField(
+        required=False, label=_('Maximum Stock'), default=0
     )
 
     image = InvenTree.serializers.InvenTreeImageSerializerField(