+++ src/frontend/src/components/plugins/PluginContext.tsx
@@ -48,7 +48,9 @@ import {
 import { useServerApiState } from '../../states/ServerApiState';
 import { InvenTreeTableInternal } from '../../tables/InvenTreeTable';
 import { EditApiForm } from '../forms/ApiForm';
+import { Thumbnail } from '../images/Thumbnail';
 import { RenderInstance, RenderRemoteInstance } from '../render/Instance';
+import { RenderInlineModel } from '../render/Instance';
 
 export const useInvenTreeContext = () => {
   const [locale, host] = useLocalState(useShallow((s) => [s.language, s.host]));
@@ -82,6 +84,8 @@ export const useInvenTreeContext = () => {
       useInstance: useInstance,
       renderInstance: RenderInstance,
       renderRemoteInstance: RenderRemoteInstance,
+      renderInlineModel: RenderInlineModel,
+      thumbnail: Thumbnail,
       theme: theme,
       colorScheme: colorScheme,
       importer: {